+++ src/functions/arithmetic/add.js
@@ -12,9 +12,9 @@ class LimeFunctionAdd extends LimeFunction {
 
     // Binary operation
     this.operations.b = [
-      'tz(_+)',
-      'tr(_+)',
-      'tr(_-)',
+      'tz(+_)',
+      'tr(+_)',
+      'tr(-_)',
 
       'cb(int->rat,rat)',
       'cb(rat,int->rat)',
@@ -29,8 +29,8 @@ class LimeFunctionAdd extends LimeFunction {
     });
     this.algorithms.set('b(rat,rat)', (step) => {
       step.bs(this.lime.build('rational')(
-        this.lime.build('integer')(step.left.value.n * step.right.value.d + step.right.value.n * step.left.value.d),
-        this.lime.build('integer')(step.left.value.d * step.right.value.d),
+        this.lime.direct([step.left.nPlace, '*', step.right.dPlace, '+', step.right.nPlace, '*', step.left.dPlace]),
+        this.lime.direct([step.left.dPlace, '*', step.right.dPlace]),
       ));
     });
   }