+++ src/functions/arithmetic/positive.js
@@ -12,18 +12,18 @@ class LimeFunctionPositive extends LimeFunction {
 
     // Right unary operation
     this.operations.r = [
-      'tr(_+)',
-      'tr(_-)',
+      'tr(+_)',
+      'tr(-_)',
 
       'er(int)',
       'er(rat)',
-      'er(_+)',
-      'er(_-)',
+      'er(+_)',
+      'er(-_)',
     ];
 
     // Algorithms
     this.algorithms.set('r(int)', (step) => {
-      step.rus(this.lime.build('integer')(step.right.value));
+      step.rus(this.lime.direct([1, '*', step.right]));
     });
     this.algorithms.set('r(rat)', (step) => {
       step.rus(this.lime.build('rational')(
@@ -31,11 +31,11 @@ class LimeFunctionPositive extends LimeFunction {
         step.right.dPlace,
       ));
     });
-    this.algorithms.set('r(_+)', (step) => {
-      step.rus(this.lime.refer('_+'));
+    this.algorithms.set('r(+_)', (step) => {
+      step.rus(this.lime.refer('+_'));
     });
-    this.algorithms.set('r(_-)', (step) => {
-      step.rus(this.lime.refer('_-'));
+    this.algorithms.set('r(-_)', (step) => {
+      step.rus(this.lime.refer('-_'));
     });
   }
 }