+++ src/system/oplist.js
@@ -14,6 +14,9 @@ const Oplist = {
     // [Binary] Rational, Rational
     'b(rat,rat)': (step) => step.bpi('rational'),
 
+    // [Left unary] Integer
+    'l(int)': (step) => step.lpi('integer'),
+
     // Nested
     'n()': () => true,
 
@@ -27,10 +30,10 @@ const Oplist = {
     'r(+)': (step) => step.rpi('add'),
 
     // [Right unary] Negative
-    'r(_-)': (step) => step.rpi('negative'),
+    'r(-_)': (step) => step.rpi('negative'),
 
     // [Right unary] Positive
-    'r(_+)': (step) => step.rpi('positive'),
+    'r(+_)': (step) => step.rpi('positive'),
 
     // [Right unary] Subtract
     'r(-)': (step) => step.rpi('subtract'),
@@ -41,14 +44,14 @@ const Oplist = {
 
   // Actions
   act: {
-    // Transfer function from add to positive
-    'f(_+)': (step) => {
-      step.fs(step.lime.refer('_+'));
+    // Transfer function to negative
+    'f(-_)': (step) => {
+      step.fs(step.lime.refer('-_'));
     },
 
-    // Transfer function from subtract to negative
-    'f(_-)': (step) => {
-      step.fs(step.lime.refer('_-'));
+    // Transfer function to positive
+    'f(+_)': (step) => {
+      step.fs(step.lime.refer('+_'));
     },
 
     // Convert left parameter from integer to rational
@@ -61,14 +64,14 @@ const Oplist = {
       step.rps(step.right.toRational());
     },
 
-    // Transfer right parameter from add to positive
-    'r(_+)': (step) => {
-      step.rps(step.lime.refer('_+'));
+    // Transfer right parameter to negative
+    'r(-_)': (step) => {
+      step.rps(step.lime.refer('-_'));
     },
 
-    // Transfer right parameter from subtract to negative
-    'r(_-)': (step) => {
-      step.rps(step.lime.refer('_-'));
+    // Transfer right parameter to positive
+    'r(+_)': (step) => {
+      step.rps(step.lime.refer('+_'));
     },
   },
 
@@ -80,17 +83,17 @@ const Oplist = {
     // [Binary] Convert right parameter from integer to rational
     'cb(rat,int->rat)': ['b(rat,int)', 'r(int->rat)'],
 
-    // [Right unary] Transfer right parameter from add to positive
-    'tr(_+)': ['r(+)', 'r(_+)'],
+    // [Right unary] Transfer right parameter to positive
+    'tr(+_)': ['r(+)', 'r(+_)'],
 
-    // [Right unary] Transfer right parameter from subtract to negative
-    'tr(_-)': ['r(-)', 'r(_-)'],
+    // [Right unary] Transfer right parameter to negative
+    'tr(-_)': ['r(-)', 'r(-_)'],
 
-    // Transfer at zero position from add to positive
-    'tz(_+)': ['z()', 'f(_+)'],
+    // Transfer at zero position to positive
+    'tz(+_)': ['z()', 'f(+_)'],
 
-    // Transfer at zero position from subtract to negative
-    'tz(_-)': ['z()', 'f(_-)'],
+    // Transfer at zero position to negative
+    'tz(-_)': ['z()', 'f(-_)'],
   },
 };